# Pop intersected postings heap without popping