@colega 2021-12-29
These are the benchmarks compared to pre-#9907 main:
```
name                                                                   old time/op    new time/op    delta
Querier/Head/labelValuesWithMatchers/i_with_n="1"                         140ms ± 0%     101ms ± 1%   -27.70%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"                      1.70s ± 7%     0.17s ± 1%   -89.95%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"                130ms ± 1%     116ms ± 1%   -10.33%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"      274ms ± 1%     171ms ± 2%   -37.64%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"                      701ms ± 2%     137ms ± 0%   -80.46%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/n_with_i="1"                        5.48µs ± 4%    4.54µs ± 2%   -17.17%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1"                        111ms ±13%     111ms ± 0%      ~     (p=0.151 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"                     1.34s ± 3%     0.14s ± 0%   -89.26%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"              86.2ms ± 7%   116.9ms ± 1%   +35.52%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"     177ms ± 3%     162ms ± 0%    -8.54%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"                     1.15s ± 2%     0.15s ± 0%   -86.72%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/n_with_i="1"                        906µs ± 8%     928µs ± 1%      ~     (p=0.151 n=5+5)

name                                                                   old alloc/op   new alloc/op   delta
Querier/Head/labelValuesWithMatchers/i_with_n="1"                        29.5MB ± 0%    17.5MB ± 0%   -40.77%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"                     46.8MB ± 0%    34.7MB ± 0%   -25.71%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"               29.5MB ± 0%    17.5MB ± 0%   -40.79%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"     46.8MB ± 0%    19.1MB ± 0%   -59.20%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"                     10.5kB ± 2%     5.7kB ± 0%   -45.57%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/n_with_i="1"                        5.54kB ± 0%    4.55kB ± 0%   -17.89%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1"                       39.1MB ± 0%    19.1MB ± 0%   -51.23%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"                     287MB ± 0%      36MB ± 0%   -87.33%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"              34.3MB ± 0%    19.1MB ± 0%   -44.40%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"    51.6MB ± 0%    20.7MB ± 0%   -59.90%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"                     144MB ± 0%       0MB ± 0%      ~     (p=0.079 n=4+5)
Querier/Block/labelValuesWithMatchers/n_with_i="1"                       6.43kB ± 0%    6.12kB ± 0%    -4.85%  (p=0.008 n=5+5)

name                                                                   old allocs/op  new allocs/op  delta
Querier/Head/labelValuesWithMatchers/i_with_n="1"                          104k ± 0%      200k ± 0%   +92.40%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"                       204k ± 0%      300k ± 0%   +47.10%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"                 104k ± 0%      200k ± 0%   +92.35%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"       204k ± 0%      200k ± 0%    -1.94%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"                       67.0 ± 1%      99.0 ± 0%   +47.76%  (p=0.008 n=5+5)
Querier/Head/labelValuesWithMatchers/n_with_i="1"                          61.0 ± 0%      87.0 ± 0%   +42.62%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1"                         304k ± 0%      300k ± 0%    -1.30%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"                     5.20M ± 0%     0.40M ± 0%   -92.31%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"                204k ± 0%      300k ± 0%   +47.09%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"      304k ± 0%      300k ± 0%    -1.31%  (p=0.008 n=5+5)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"                     3.00M ± 0%     0.00M ± 0%      ~     (p=0.079 n=4+5)
Querier/Block/labelValuesWithMatchers/n_with_i="1"                         61.0 ± 0%     129.0 ± 0%  +111.48%  (p=0.008 n=5+5)
```

There's still an increase in allocations, but now it's at most 2x.